ICC fixes.

+++ tests/server.c
@@ -61,7 +61,7 @@ void server_startup(server_startup_st *construct)
           {
             if (fgets(buffer, sizeof(buffer), fp) != NULL)
             { 
-              pid_t pid = (pid_t)atol(buffer);
+              pid_t pid= (pid_t)atoi(buffer);
               if (pid != 0) 
                 kill(pid, SIGTERM);
             }
@@ -99,7 +99,7 @@ void server_startup(server_startup_st *construct)
 
   for (x= 0; x < memcached_server_list_count(construct->servers); x++)
   {
-    printf("\t%s : %u\n", construct->servers[x].hostname, construct->servers[x].port);
+    printf("\t%s : %d\n", construct->servers[x].hostname, construct->servers[x].port);
     assert(construct->servers[x].fd == -1);
     assert(construct->servers[x].cursor_active == 0);
   }
